requirements.txt
cloudpickle>=3.1.2
contourpy>=1.3.3
cycler>=0.12.1
fonttools>=4.61.1
ImageIO>=2.37.2
imbalanced-learn>=0.14.1
joblib>=1.5.3
kiwisolver>=1.4.9
lazy_loader>=0.4
lime>=0.2.0.1
llvmlite>=0.46.0
matplotlib>=3.10.0
networkx>=3.0
numba>=0.60.0
numpy>=1.26.0
pandas>=2.1.0
pillow>=10.0.0
pyparsing>=3.0.0
python-dateutil>=2.8.2
scikit-image>=0.22.0
scikit-learn>=1.4.0
scipy>=1.12.0
seaborn>=0.13.0
shap>=0.44.0
six>=1.16.0
sklearn-compat>=0.1.5
slicer>=0.0.8
threadpoolctl>=3.2.0
tifffile>=2023.0.0
tqdm>=4.66.0
typing_extensions>=4.9.0
xgboost>=2.0.0
streamlit>=1.35.0
watchdog>=4.0.0
plotly>=5.20.0

# ── QML Research Dependencies ────────────────────────────────────────────────
pennylane>=0.36.0          # Quantum computing framework (PQC / QNN / Kernels)
torch>=2.0.0               # PyTorch backend for QNN gradient computation

# ── XAI Expansion Dependencies ───────────────────────────────────────────────
alibi>=0.9.0               # Anchor explanations (rule-based XAI)
dice-ml>=0.11              # DiCE counterfactual explanations
